separate functions from main package
